more cleanup and text changes
Removed color codes from console messages as they really aren't needed. Changes variable prefixes for console messages to text. While I wanted Lock-IT to have similar features+ that Lockette has, I didn't want it to "mirror" the plugin. With that being said, I remove all traces of [Private] and replaced it with [Lock-IT] for the protection signs.